crypto: Create method for parsing string of cipher suites
This is a utility function that allows you to parse a set of cipher
suites, and get a slice containing the const values for each cipher
suite.

The main purpose of this is to be able to easily parse such
configurations for HTTP(s) servers without much boiler plate code.

Fixes: #30325

Change-Id: Icdd564e0fe9b605862c7cdbb33a5d9de9af8f50a
GitHub-Last-Rev: f01fc499de002d16edbe648deb9223c44fc25d70
GitHub-Pull-Request: golang/go#30326