Update patch set 1
Patch Set 1: Code-Review-1

In general the churn here on the code base does not seem worth it.

Removing trailing spaces in the html files does not look right either.
Also changes in the vendored directories need to be upstreamed first then vendored back in which if made at all should be their own CLs.

If this is a general style guideline then it would be good if tests and/or gofmt prevent regressions in the future.

Patch-set: 1
Reviewer: Gerrit User 16006 <16006@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-1