cmd/compile: use correct NOP on ppc64le for stacktraces with inlines
A recent change to fix stacktraces for inlined functions,
introduced a regression on ppc64le when compiling position
independent code. That happened because ginsnop2 was used
for this new purpose when it should use ginsnop to insert a
NOP to identify the inlined function location.

This fixes that problem.

Fixes #30283

Change-Id: If70ef27fc65ef31969712422306ac3a57adbd5b6