cmd/compile: use correct NOP on ppc64le for stacktraces with inlines
A change to fix stacktraces for inlined functions caused a
regression when building PIC on ppc64le. In that change 69c2c56,
a single NOP should be inserted at the point where a function
has been inlined to allow for correct stacktraces. Instead, a
function was used on ppc64le that is really only intended for a
special case when inserting instructions before a deferreturn,
where the load of r2 is needed when compiling for PIC.

This fixes that problem.

Fixes #30283

Change-Id: If70ef27fc65ef31969712422306ac3a57adbd5b6