Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Patch Set 3:
> > > 
> > > > Patch Set 3: Code-Review+2
> > > > 
> > > > LGTM. Is it possible to have a test?
> > > 
> > > Yes. Actually just more testing of the options that set the shared flags (-buildmode=plugin, -buildmode=pie) would have caught this. But I will look at what can be added to the existing tests for these buildmodes to force an error situation.
> > 
> > After looking at this a bit, I'm not sure a testcase can consistently reproduce the problem. We mainly need more testing of the buildmodes that generate position independent code.
> 
> Does misc/cgo/testcshared or testplugin work? Maybe call a function in another module in a defer statement?
> 
> We can probably check this in first, and add a test later.

Yes, let's add a testcase later. It's not defer that causes the issue, it's when the load of r2 is done when it is not a defer. Because in that case it could load a value that is not initialized. The deferreturn case has always worked and my patch doesn't change that behavior.

Patch-set: 3
Reviewer: Gerrit User 6320 <6320@62eb7196-b449-3ce5-99f1-c037f21e1705>