Update patch set 4
Patch Set 4:

(1 comment)

I was talking to Austin and he suggested that the way forward was to monitor usage during the last GC cycle and only release objects above the high water mark. The implementation is similar to the code here in the sense the pool is split in two. One can think the split as a main cache and a victim cache. The main cache is used to supply objects. When the main cache is empty objects come from the victim cache. When both are empty new objects are allocated. Objects are always returned to the main cache. When a GC cycle starts the victim cache releases all its objects, the main cache becomes the victim cache, and the main cache starts out empty.

Floating garbage is limited to 1 GC cycle but the number of objects released at each GC reflects the high water mark of the most recent GC cycle instead of the state at the start of a GC cycle. This seems more robust than simply releasing half the objects.

Patch-set: 4
Reviewer: Gerrit User 5186 <5186@62eb7196-b449-3ce5-99f1-c037f21e1705>