Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> I believe this change broke the iOS builders.

Yep. I think it triggered a bug in the iOS exec script: it's comparing the working directory to the GOPATH directory, but GOPATH is set relative to TMPDIR, TMPDIR happens to be a symlink, and os.Getwd ends up picking up the resolved path instead.

The more robust solution is probably to fix the exec script to normalize symlinks (see cmd/go/internal/search.InDir), but in the interim it may suffice to set PWD explicitly.

Patch-set: 9