runtime: scan defer closure in stack scan
With stack objects, when we scan the stack, it scans defers with
tracebackdefers, but it seems to me that tracebackdefers doesn't
include the func value itself, which in this case is a stack
allocated closure. Scan it explicitly.

Alternatively, we can change tracebackdefers to include the func
value, which in turn needs to change the type of stkframe.

TODO: need test.

Change-Id: I55a6e43264d6952ab2fa5c638bebb89fdc410e2b