Update patch set 4
Patch Set 4:

> Patch Set 3: Code-Review+2
> 
> Code itself looks fine.
> 
> I'm not terribly happy with the naming. The fact that ginsnop2 does more than insert nops is confusing, and what tricked me into introducing this bug in the first place.
> Maybe rename Ginsnopdefer to something else?
> Maybe always call Ginsnop, then also call PPCOnlyDeferreturnPreamble which does the R2 load and the second nop? (All other archs this call would be a nop.)
> 
> I'll +2 for now because it's a nice minimal change for 1.12, but I'd like to fix the naming for 1.13.

Sorry I was on vacation for a few days... I did not put ginsnop2 in here originally and I'm not sure why there is a need for 2 instructions. I understand why in some cases the ld r2 needs to be there when doing the deferreturn, but I think it could be either nop or ld r2 not both. But I'd have to spend some time to test it out and understand what's happening.

Patch-set: 4