cmd/link: convert sym.Symbol to use encoded/interned name
DO NOT REVIEW
DO NOT SUBMIT

Switch over sym.Symbol to use encoded/interned names instead
of raw strings for the "name" and "extname" fields.

Notes:

This change slows down the linker considerably; a good deal of
additional work will be needed to recover the performance. The main
problem here is that there are many places in the linker where we have
a loop over ctxt.Syms.Allsym that look for symbols with particular
names or prefixes (ex: "type.*" or "_cgo_..."). Each of these loops
now has to allocate a string in order to materialize the sym name.

This seems fixable by exploring things like symbol flags (e.g.
set a bit in the flags if a given symbol is a "type" symbol) and/or
better uses of symbol names (for example, prefix/suffix-aware version
versions of HasPrefix/HasSuffix).

Change-Id: I824734912b555d1b4129921f754f2f0d057436c2