Update patch set 1
Patch Set 1:

> $ diff old new
> 5,6d5
> < ./decode.go:335:41: Found IsInBounds
> < ./decode.go:348:26: Found IsInBounds

Err, I misread the compiler debug output. IsInBounds is a bounds check, not an assertion that it removed a bounds check. So this CL is still removing the bounds checks, as expected.

But benchstat can no longer show a noticeable speedup. I might give that another go later.

Patch-set: 1