Update patch set 4
Patch Set 4:

> I was talking to Austin and he suggested that the way forward was to monitor usage during the last GC cycle and only release objects above the high water mark. The implementation is similar to the code here in the sense the pool is split in two. One can think the split as a main cache and a victim cache. The main cache is used to supply objects. When the main cache is empty objects come from the victim cache. When both are empty new objects are allocated. Objects are always returned to the main cache. When a GC cycle starts the victim cache releases all its objects, the main cache becomes the victim cache, and the main cache starts out empty.

Such a design would have the benefit of further reducing the number of allocations, as it's deliberately retaining more objects alive, but it's otherwise pretty similar: just consider the odd/even poolLocals as alternating main/victim caches, and have objects put randomly in both main and victim caches (instead of always in the main, as proposed, that is why the proposed design retains strictly more objects alive).

As was mentioned, though, it's probably a good idea to be conservative and err on the side of retaining less alive than err on the side of retaining more.

> This seems more robust than simply releasing half the objects.

"Robust" in what sense? Such a design sounds to me more subsceptible to spiky allocation patterns (e.g. with the proposed design if during a cycle an abnormal number of objects are returned to the pool, in the next cycle they will *all* be unconditionally retained; for this to happen with the code in this CL all objects should be returned by the same P *and* its poolLocal should not be cleared during that cycle). This CL, as it's more conservative, should be less suscebitble to that scenario.

Also note that this CL does not clear half the *objects*. It alternatively clears half of the *poolLocals* (as a bonus, this makes poolCleanup slightly faster).

I believe this CL would be a good first step in alleviate the problem described in the gihub issue, as it's more conservative than the alternative design. If then evidence shows that it's not enough I am all for more aggressive options.

Patch-set: 4
Reviewer: Gerrit User 11500 <11500@62eb7196-b449-3ce5-99f1-c037f21e1705>