Update patch set 5
Patch Set 5:

Joe, Brad, Ian, any strong opinions on how exactly to change the error type?

This could technically break a program that depends on Field just being the field name, but I think that's unlikely. And it would have a simple fix too. We can always merge this and document it in the changelog, and see if it causes any trouble.

Patch-set: 5
Reviewer: Gerrit User 13550 <13550@62eb7196-b449-3ce5-99f1-c037f21e1705>