Update patch set 7
Patch Set 7:

(1 comment)

Thanks! Code looks good.
I just think the commit message could be clearer on what
is relevant for the end user (parsing {Set,}Cookie headers)
and what is the technical rational behind the change.
The fact that readCookies is changed is not that important
and easily to determine from the diff.

Patch-set: 7