cmd/compile: fix ordering for short-circuiting ops
Make sure the side effects inside short-circuited operations (&& and ||)
happen correctly.

Before this CL, we attached the side effects to the node itself using
exprInPlace. That caused other side effects in sibling expressions
to get reordered with respect to the short circuit side effect.

Instead, rewrite a && b like:

r := a
if r {
  r = b
}

That code we can keep correctly ordered with respect to other
side-effects extracted from part of a big expression.

exprInPlace seems generally unsafe. But this was the only case where
exprInPlace is called not at the top level of an expression, so I
don't think the other uses can actually trigger an issue (there can't
be a sibling expression). TODO: maybe those cases don't need "in
place", and we can retire that function generally.

This CL needed a small tweak to the SSA generation of OIF so that the
short circuit optimization still triggers. It can handle triangles but
not diamonds, so if the "then" branch is empty don't generate a block
for it.

TODO: decide whether changes to test/*.go make sense or not.
TODO: benchmark
TODO: remove OANDAND/OOROR cases from walk? order now removes them
     (but it isn't obvious, because walk reintroduces some.)

Fixes #30566

Change-Id: I19c04296bea63cbd6ad05f87a63b005029123610