Update patch set 1
Patch Set 1:

Hello Vladimir, thank you for this change!

As Ian mentioned, this change can't fly currently and also
packages such as runtime can be readily navigated to at /runtime
when reading this code even on a browser, thus I am going to abandon this CL.

We do appreciate the contribution though and look forward to even more.

Patch-set: 1
CC: Gerrit User 5137 <5137@62eb7196-b449-3ce5-99f1-c037f21e1705>