cmd/go: improve wording of 'go mod init' error
When 'go mod init' is run without a module path, it tries to infer a
module path, based on the current directory (if in GOPATH), import
comments, and vendor configuration files.

It's common for this command to fail the first time a user tries to
create a module in a new project outside GOPATH. This change improves
the wording of the error message to hint that the user should specify
a module path.

Change-Id: Iec0352e919dbc8b426ab71eed236fad3929ec671