net/http: expose standard transport for js
Implement a separate type jsTransport that implements RoundTripper using
the WHATWG Fetch API, and make it the DefaultTransport on js/wasm. This
allows using the standard Transport implementation on js/wasm by
specifying a DialContext for the Transport. Note that the default would
use net.Dial, that can not be used in the browser.

Before this change the same Transport type was used also on js, but with
a different implementation. This implementation disregarded any of the
documented Transport settings and just used the browser Fetch API. This
made it impossible to use the Transport implementation, e.g., by
supplying a DialContext that works in the browser.

Fixes #27495

Change-Id: Ia0876890c2562fe48bea662c278ba951ac884767