Update patch set 19
Patch Set 19:

This change seem to have broken the wasm builder:

https://build.golang.org/log/daec1a0afc3ad8c6fd44563e17aafa961a6e2552

I think the same error will appear on the iOS builder. Likewise, recent Android builds complain about missing git (the android exec wrapper installs the go tool on the device but not git).

Also the subrepos are a sea of red. I don't know if this or a follow-up change did that.

Patch-set: 19
CC: Gerrit User 30809 <30809@62eb7196-b449-3ce5-99f1-c037f21e1705>