cmd/go/internal/load: use path.Split on import path in DefaultExecName
This change is a follow-up to CL 140863. That change factored out
the logic to determine the default executable name from the Package.load
method into a DefaultExecName function, and started using it in more
places to avoid having to re-implement the logic everywhere it's needed.

Most previous callers already computed the default executable name based
on the import path. The load.Package method, before CL 140863, was the
exception, in that it used the p.Dir value in GOPATH mode instead.
There was a comment that it should be equivalent to use import path,
but it was too late in Go 1.11 cycle to risk implementing that change.
We're relatively early in Go 1.12 cycle now, so now is a better time
to make it happen for Go 1.13.

Refactor DefaultExecName to be a pure function that takes two inputs
that it needs. This makes it clear that it doesn't need most other
fields from the Package struct. Also document the algorithm for how
the default executable name is determined in DefaultExecName.

Fixes golang/go#30821
Updates golang/go#26869

Change-Id: Ib1ebb95acba7c85c24e3a55c40cdf48405af34f3