Update patch set 9
Patch Set 9:

> Like I said before, I do not know if that difference between Unix and Windows implementations is important. So leaving for Ian to decide what to do here.

The difference is between a blocking-mode Unix fd and others.
Before this change, Windows, Unix non-blocking-mode fd and Unix blocking-mode fd had three different behaviors. With this change, there are only two.

> Otherwise, if you remove fd.fdmu.closed calls, it LGTM. I am not convinced by your "people will call CancelIoEx at random moments" argument.

It's OK to remove that call and the result is almost the same.
But my opinion is that it's not correct to return ErrClosed before checking whether the file is really closed.

Patch-set: 9
Reviewer: Gerrit User 22895 <22895@62eb7196-b449-3ce5-99f1-c037f21e1705>