Update patch set 3
Patch Set 3:

(4 comments)

Cool, thanks for working on this!
Minor nits about comments and unrealistic "identity" usages.

One bigger comment on the structure:
It seems the code interprets TE header twice: 
- in fixTransferEncoding() with a single pass over the list
- in readTransfer() by multiple "global" questions like chunked() || implicitlyChunked(), gzipped(), etc.
The effective logic seems identical (identity corner cases aside), but it's hard to be sure.
Could readTransfer first do fixTransferEncoding() and then use it in simpler ways?
Hmm, it already calls fixTransferEncoding() 👍
I think the missing part for some cleanup is for fixTransferEncoding() to translate implicitly chunked to explicitly chunked.

Then, some helpers like chunked() could work at a "higher layer" assuming the input is already sanitized.  It is currently hard to see which parts work on original TE header and which on the  sanitized; making that distinction explicit would IMO be helpful to readers.

P.S. I'm not at all an expert on TE or HTTP standards.

Patch-set: 3
CC: Gerrit User 31400 <31400@62eb7196-b449-3ce5-99f1-c037f21e1705>