cmd/link/internal/ld: copy LC_VERSION_MIN_* commands to Go object file
To run on a non-default platform (macOS for 386 and amd64, iOS for
arm/arm64), the Apple toolchain requires a Mach-O load command that
describes the platform and minimum version for all object files in
a build. The go.o object file produced by the Go linker contains no
such command.
The loader commands are mutually exclusive so we need to pick the
right one. Fortunately, the command is only really needed for watchOS
and tvOS where cgo is always enabled, which means we can copy the
first loader command we find in the object files produced by the
host compiler.

Updates #22395

Change-Id: I7a47d19be9d80f0459dc358c600cddd9f236c444