Revert "cmd/go: fix the default build output name for versioned binaries"
DO NOT SUBMIT: wait for me to send the improved fix and get that reviewed;
               then both can be submitted at once. no need to submit this
               earlier than that.

This reverts CL 140863 (commit bf94fc3ae387fc09929443393741919fac6727af).

Reason for revert: There was a potential problem spotted in the original
fix, which resulted in it being rolled back from release-branch.go1.12
and not included in Go 1.12.1 release. We intend to improve the fix and
include it in Go 1.12.2 instead. To make the fix easier to backport,
revert this change before applying the improved fix.

Change-Id: If6c785f58482d2531b5927c5ea7002f548c21c7c