cmd/link: add optional sanity checking for DWARF symbols
Introduce a new linker command line option "-strictdwarf", which
enables sanity checking of duplicate DWARF info symbols.

Currently if we read a DWARF symbol (such as "go.info.PKG.FUNCTION") from one
object file, and then encounter the same symbol later on while reading
another object file, we simply discard the second one and move on
with the link, since the two should be (in theory) be identical.

If as a result of a compiler bug we wind up with symbols that are
not identical, this tends to (silently) result in bad/incorrect DWARF
generation, which may or may not be discovered depending on who is
consuming the DWARF and what't being done with it.

When this option is turned on, at the point where a duplicate DWARF
symbol is detected in the object file reader, we check to make sure
that the length/contents of the symbol are the same as the previously
read symbol, and print a descriptive warning if not.

This option should provide an easier way for flushing out these sorts
of problems, and might hopefully enable us to (at some point) turn on
stricter DWARF checks in the linker by default.

[NB: not sure if it makes sense to submit this or not.]

Updates #30908.

Change-Id: I64c4e07c326b4572db674ff17c93307e2eec607c