Update patch set 2
Patch Set 2:

I think it would be good to have this check handy, and the cost looks low.
Any chance of a test, or is that likely to be a pain to construct and maintain?
(Or, have you used this at least once on flaky DWARF to verify that the error messages look sensible?)

Patch-set: 2
Reviewer: Gerrit User 7061 <7061@62eb7196-b449-3ce5-99f1-c037f21e1705>